+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to `peppool-wallet` are documented in this file.
 
 ## [Unreleased]
 
+### Fixed
+- Show net amount for outgoing transactions ([#15](https://github.com/mvdnbrk/peppool-wallet/pull/15))
+
 ## [0.2.0](https://github.com/mvdnbrk/peppool-wallet/releases/tag/v0.2.0) - 2026-04-08
 
 ### Added